third_party/WebKit/LayoutTests/external/wpt/service-workers/service-worker/registration-updateviacache.https-expected.txt

This is a testharness.js-based test.
FAIL register-via-api-updateViaCache-undefinedassert_equals: reg.updateViaCache expected (string) "imports" but got (undefined) undefined
FAIL register-via-api-updateViaCache-importsassert_equals: reg.updateViaCache expected (string) "imports" but got (undefined) undefined
FAIL register-via-api-updateViaCache-allassert_equals: reg.updateViaCache expected (string) "all" but got (undefined) undefined
FAIL register-via-api-updateViaCache-noneassert_equals: reg.updateViaCache expected (string) "none" but got (undefined) undefined
FAIL register-via-link-element-updateViaCache-undefinedassert_equals: reg.updateViaCache expected (string) "imports" but got (undefined) undefined
FAIL register-via-link-element-updateViaCache-importsassert_equals: reg.updateViaCache expected (string) "imports" but got (undefined) undefined
FAIL register-via-link-element-updateViaCache-allassert_equals: reg.updateViaCache expected (string) "all" but got (undefined) undefined
FAIL register-via-link-element-updateViaCache-noneassert_equals: reg.updateViaCache expected (string) "none" but got (undefined) undefined
FAIL register-via-link-header-updateViaCache-undefinedassert_equals: reg.updateViaCache expected (string) "imports" but got (undefined) undefined
FAIL register-via-link-header-updateViaCache-importsassert_equals: reg.updateViaCache expected (string) "imports" but got (undefined) undefined
FAIL register-via-link-header-updateViaCache-allassert_equals: reg.updateViaCache expected (string) "all" but got (undefined) undefined
FAIL register-via-link-header-updateViaCache-noneassert_equals: reg.updateViaCache expected (string) "none" but got (undefined) undefined
FAIL register-with-updateViaCache-undefined-then-undefinedassert_equals: reg.updateViaCache updated expected (string) "imports" but got (undefined) undefined
FAIL register-with-updateViaCache-undefined-then-importsassert_equals: reg.updateViaCache updated expected (string) "imports" but got (undefined) undefined
FAIL register-with-updateViaCache-undefined-then-allassert_equals: reg.updateViaCache updated expected (string) "all" but got (undefined) undefined
FAIL register-with-updateViaCache-undefined-then-noneassert_equals: reg.updateViaCache updated expected (string) "none" but got (undefined) undefined
FAIL register-with-updateViaCache-imports-then-undefinedassert_equals: reg.updateViaCache updated expected (string) "imports" but got (undefined) undefined
FAIL register-with-updateViaCache-imports-then-importsassert_equals: reg.updateViaCache updated expected (string) "imports" but got (undefined) undefined
FAIL register-with-updateViaCache-imports-then-allassert_equals: reg.updateViaCache updated expected (string) "all" but got (undefined) undefined
FAIL register-with-updateViaCache-imports-then-noneassert_equals: reg.updateViaCache updated expected (string) "none" but got (undefined) undefined
FAIL register-with-updateViaCache-all-then-undefinedassert_equals: reg.updateViaCache updated expected (string) "imports" but got (undefined) undefined
FAIL register-with-updateViaCache-all-then-importsassert_equals: reg.updateViaCache updated expected (string) "imports" but got (undefined) undefined
FAIL register-with-updateViaCache-all-then-allassert_equals: reg.updateViaCache updated expected (string) "all" but got (undefined) undefined
FAIL register-with-updateViaCache-all-then-noneassert_equals: reg.updateViaCache updated expected (string) "none" but got (undefined) undefined
FAIL register-with-updateViaCache-none-then-undefinedassert_equals: reg.updateViaCache updated expected (string) "imports" but got (undefined) undefined
FAIL register-with-updateViaCache-none-then-importsassert_equals: reg.updateViaCache updated expected (string) "imports" but got (undefined) undefined
FAIL register-with-updateViaCache-none-then-allassert_equals: reg.updateViaCache updated expected (string) "all" but got (undefined) undefined
FAIL register-with-updateViaCache-none-then-noneassert_equals: reg.updateViaCache updated expected (string) "none" but got (undefined) undefined
Harness: the test ran to completion.
